+++ src/main/index.js
@@ -1,6 +1,8 @@
 import './globalSetting'
-
+import { checkSystem } from './utils/checkSystem'
 import App from './app'
+
+checkSystem()
 
 const app = new App()